Move test functions to utils test package
